more updates from kenneth
git-svn-id: http://zhihaolou.org/svn/neoParSA/trunk@288 a3e5a195-83b0-4c6b-8637-e6d081586ea5

+++ tests/transcpp/score.h
@@ -41,7 +41,7 @@ private:
   
   vector<double> scores;
   vector<double> weights;
-  int min_weight;
+  double min_weight;
   
   double score;
